+++ sharedJS/entity.mjs
@@ -4,14 +4,18 @@ import { Circle } from "./shapes.mjs";
 
 let idGen = 0;
 class Entity {
-    constructor(location, imgSrc) {
+    constructor(location, imgSrc, hitbox, lookDirection = new Vec2(1,0), speed = 0) {
         if(!(location instanceof Vec2)) {
             throw TypeError("Entity: Location not Vec2");
         }
         this.id = idGen++;
         this.location = location;
-        this.image = new Image();
-        this.image.src = imgSrc;
+        this.oldLocation = location;
+        //this.image = new Image();
+        this.imgSrc = imgSrc;
+        this.hitbox = hitbox;
+        this.lookDirection = lookDirection;
+        this.speed = speed;
         //mostly for debugging now
         this.overlapping = false;
     }
@@ -32,38 +36,29 @@ class Entity {
         let objectJson = JSON.stringify(this);
         return {type: objectType, json: objectJson};
     }
-}
-
-class Moveable extends Entity {
-    constructor(location, imgSrc, speed, lookDirection) {
-        super(location,imgSrc);
-        this.speed = speed;
-        this.moved = true; //true so it hopefully gets drawn frame it spawns
-        //if lookDirection isn't set default to looking left
-        this.lookDirection = lookDirection || new Vec2(1,0);
-    }
     update(dt, map) {
-        this.move(dt, this.lookDirection, map);
+        if(this.speed > 0) {
+            this.move(dt, this.lookDirection, map);
+        }
     }
     move(dt, direction, map) {
         if(!(direction instanceof Vec2)) {
             throw TypeError("Moveable move: Direction not Vec2");
         }
-        let oldLocation = this.location.clone();
+        this.oldLocation = this.location.clone();
 
         let dist = this.speed * dt;
         this.location.addS(direction.getUnit().multiplyScalar(dist));
+        /*
+        //handle this with collision detection
         if(this.location.x < 0 || this.location.x >= map.canvas.width) {
             this.location.x = oldLocation.x;
         }
         if(this.location.y < 0 || this.location.y >= map.canvas.height) {
             this.location.y = oldLocation.y;
         }
-    }
-    draw(map) {
-        //map.drawImageLookat(this.image, this.location, this.lookDirection);
-        map.drawImageLookat(this.image, this.location, this.lookDirection, this.overlapping);
+        */
     }
 }
 
-export { Entity, Moveable };
+export default Entity;
