+++ views/game.ejs
@@ -3,7 +3,7 @@
 
 <canvas id="game"></canvas>
 
-<script src="/socket.io/socket.io.js"></script>
+<script src="js/socket.io/socket.io.js"></script>
 <script async type="text/javascript" src="js/keyBinds.js"></script>
 <script async type="module" src="js/gameLoop.js"></script>
